+++ examples/data-sources/thundercompute_gpu_availability/data-source.tf
@@ -1,0 +1,5 @@
+data "thundercompute_gpu_availability" "current" {}
+
+output "gpu_availability" {
+  value = data.thundercompute_gpu_availability.current.specs
+}